requirements.txt
Django==5.1.3
djangorestframework==3.15.2
google-generativeai==0.8.3
python-dotenv==1.0.1
Pillow==10.2.0
pytest==8.0.2
pytest-django==4.8.0
coverage==7.4.3
djongo==1.3.6
pymongo==3.12.3
dnspython==2.4.2
cryptography==42.0.5
streamlit==1.31.1
pandas==2.2.0
plotly==5.18.0
